requirements.txt
absl-py==1.4.0
accelerate==0.16.0
amqp==5.1.1
asttokens==2.2.1
autoflake==1.7.7
bertopic==0.13.0
black==23.1.0
brotlipy==0.7.0
cachelib==0.10.2
cachetools==5.3.0
catalyst==22.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama==0.4.6
conda==22.11.1
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1665612644927/work
cycler==0.11.0
Cython==0.29.33
dacite==1.8.0
dateparser==1.1.7
dill==0.3.6
docx2txt==0.8
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.9.0
Flask==2.0.1
Flask-Session==0.4.0
fonttools==4.39.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.3
hdbscan==0.8.29
huggingface-hub==0.13.0
hydra-slayer==0.4.1
icecream==2.1.3
idna @ file:///croot/idna_1666125576474/work
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mmh3==3.0.0
more-itertools==9.1.0
multiprocess==0.70.14
mypy-extensions==1.0.0
nltk==3.8.1
nnsplit==0.2.2
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
pathos==0.3.0
pathspec==0.11.0
pika==1.3.2
Pillow==9.4.0
platformdirs==3.1.0
plotly==5.13.1
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
polars==0.15.18
pox==0.3.2
ppft==1.7.6.6
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclean==2.6.0
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyflakes==3.0.1
Pygments==2.14.0
pynndescent==0.5.8
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
random-name==0.1.1
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
scikit-learn==1.2.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
simplejson==3.18.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==2.0.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch==1.13.1
torchvision==0.14.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
transformers==4.25.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.3
urllib3 @ file:///croot/urllib3_1670526988650/work
vine==5.0.0
Werkzeug==2.2.3
wordcloud==1.8.2.2
XlsxWriter==3.0.7
